# service/api/openapi.yaml
openapi: 3.0.3
info:
  title: Notifique API
  description: A simple notification system.
  version: 0.1.0
servers:
  - url: http://localhost:8080
    description: Local development server
paths:
...
  /users/me/notifications:
    get:
      tags:
        - users
      summary: Retrieve a page of user notifications.
      parameters:
        - $ref: "#/components/parameters/nextTokenParam"
        - $ref: "#/components/parameters/maxResultsParam"
        - in: query
          name: topic
          required: false
          description: filter notifications based on topic
          schema:
            type: array
            items:
              type: string
              minLength: 1
          explode: true
          style: form
      security:
        - OAuth2:
          - notifications/user
      responses:
        "200":
          headers:
            X-RateLimit-Limit:
              $ref: '#/components/schemas/RateLimitLimit'
            X-RateLimit-Remaining:  
              $ref: '#/components/schemas/RateLimitRemaining'
            X-RateLimit-Reset:
              $ref: '#/components/schemas/RateLimitReset'
          description: A page of notifications has been generated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PageResponseModel"
                properties:
                  data:
                    items:
                      $ref: "#/components/schemas/UserNotificationModel"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    OAuth2:
      type: oauth2
      flows:
        clientCredentials:
          tokenUrl: https://your-auth-server.com/oauth2/token
          scopes:
            user_notifications/publisher: Can publish user notifications
            notifications/publisher: Can publish notifications, update notification recipients statuses,
              update notification status
        authorizationCode:
          authorizationUrl: https://your-auth-server.com/auth
          tokenUrl: https://your-auth-server.com/token
          scopes:
            notifications/user: Can retrieve user notifications, set the read at status, 
              get the user notification config, and update the user notification config
            notifications/admin: Can retrieve notifications, get notification recipients statuses,
              cancel notifications, create/update/delete notification templates,
              create/update/delete distribution lists, and get distribution lists, add/remove recipients from distribution lists
  parameters:
    nextTokenParam:
      in: query
      name: nextToken
      required: false
      schema:
        $ref: "#/components/schemas/TokenModel"
      description: the key of the last evaluated item of the page.
    maxResultsParam:
      in: query
      name: maxResults
      required: false
      schema:
        $ref: "#/components/schemas/MaxResultsModel"
      description: the maximum number of items to return on the page.
  schemas:
    RateLimitLimit:
      type: integer
      description: The maximum number of requests allowed in the time window
    RateLimitRemaining:
      type: integer
      description: The number of requests remaining in the current time window
    RateLimitReset:
      type: integer
      format: unix-timestamp
      description: The time at which the rate limit resets
    TokenModel:
      type: string
      format: byte
      nullable: true
    MaxResultsModel:
      type: integer
      minimum: 0
    PageResponseModel:
      type: object
      properties:
        nextToken:
          $ref: "#/components/schemas/TokenModel"
        prevToken:
          $ref: "#/components/schemas/TokenModel"
        resultCount:
          type: integer
          minimum: 0
          nullable: false
        data:
          type: array
          items:
            # This is a placeholder for the actual item type, which
            # will be replaced by `$ref` in the endpoints
            type: object
      required:
        - nextToken
        - prevToken
        - resultCount
        - data
...
    UserNotificationModel:
      type: object
      properties:
        id:
          type: string
          format: uuid
          nullable: false
        title:
          type: string
          nullable: false
          minLength: 1
        contents:
          type: string
          nullable: false
          minLength: 1
        createdAt:
          type: string
          format: date-time
          nullable: false
        image:
          type: string
          format: uri
          nullable: true
        readAt:
          type: string
          format: date-time
          nullable: true
        topic:
          type: string
          nullable: false
          minLength: 1
...